services/pedidos/requirements.txt
gunicorn==19.9.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask-Testing==0.7.1
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.4.0
flake8==3.7.7
coverage==4.5.3 